+++ CHANGELOG.md
@@ -24,9 +24,9 @@
   `PaginatorInterface::getCurrentPageSize()` and `OffsetPaginator::getTotalItems()` (@vjik)
 - Chg #165: Simplify `FilterInterface` and `FilterHandlerInterface` (@vjik)
 - Chg #166: Remove `EqualsEmpty` filter (@vjik)
-- Bug #173: Make `Like` filter case-sensitive (@arogachev)
-- New #173: Add `ILike` filter (case-insensitive variation of `Like` filter) (@arogachev)
 - New #176: Add `OrderHelper` (@vjik)
+- New #173, #184: Add `$caseSensitive` parameter to `Like` filter to control whether the search must be case-sensitive
+  or not (@arogachev)
 
 ## 1.0.1 January 25, 2023